Extract operating system catalog state from the deploy shell
