+++ src/Foundry.Deploy/ViewModels/MainWindowViewModel.cs
@@ -29,10 +29,6 @@ namespace Foundry.Deploy.ViewModels;
 
 public partial class MainWindowViewModel : ObservableObject, IDisposable
 {
-    private const string DefaultReleaseId = OperatingSystemSupportMatrix.DefaultReleaseId;
-    private const string DefaultLicenseChannel = "RET";
-    private const string DefaultEdition = "Pro";
-    private const string FallbackLanguageCode = "en-us";
     private const string NoneDriverPackOptionKey = "none";
     private const string MicrosoftUpdateCatalogDriverPackOptionKey = "microsoft-update-catalog";
     private const string DellDriverPackOptionKey = "oem:dell";
@@ -45,39 +41,6 @@ public partial class MainWindowViewModel : ObservableObject, IDisposable
     private const string RuntimeFolderName = "Runtime";
     private const string WinPeTransientRuntimeRoot = @"X:\Foundry\Runtime";
     private static readonly string AppVersion = ResolveAppVersion();
-    private static readonly string DefaultLanguageCode = ResolveDefaultLanguageCode();
-    private static readonly string[] RetailEditionOptions =
-    [
-        "Home",
-        "Home N",
-        "Home Single Language",
-        "Education",
-        "Education N",
-        "Pro",
-        "Pro N",
-        "Enterprise",
-        "Enterprise N"
-    ];
-    private static readonly string[] VolumeEditionOptions =
-    [
-        "Education",
-        "Education N",
-        "Pro",
-        "Pro N",
-        "Enterprise",
-        "Enterprise N"
-    ];
-    private static readonly string[] Arm64RetailEditionOptions =
-    [
-        "Home",
-        "Pro",
-        "Enterprise"
-    ];
-    private static readonly string[] Arm64VolumeEditionOptions =
-    [
-        "Pro",
-        "Enterprise"
-    ];
     private readonly IThemeService _themeService;
     private readonly IApplicationShellService _applicationShellService;
     private readonly IExpertDeployConfigurationService _expertDeployConfigurationService;
@@ -92,16 +55,10 @@ public partial class MainWindowViewModel : ObservableObject, IDisposable
     private readonly Dispatcher _dispatcher;
     private readonly DeploymentMode _resolvedDeploymentMode;
     private readonly string? _resolvedUsbCacheRuntimeRoot;
-    private readonly HashSet<string> _configuredVisibleLanguageCodes = new(StringComparer.OrdinalIgnoreCase);
     private HardwareProfile? _detectedHardware;
     private DeployMachineNamingSettings _machineNamingConfiguration = new();
-    private string? _configuredDefaultLanguageCodeOverride;
     private string _lockedComputerNamePrefix = string.Empty;
-    private bool _forceSingleVisibleLanguageSelection;
-    private bool _hasLoggedUnavailableConfiguredLanguages;
-    private bool _hasLoggedUnavailableDefaultLanguageOverride;
     private bool _isApplyingManagedComputerName;
-    private bool _isUpdatingOsFilters;
     private bool _isUpdatingDriverPackOptionSelection;
     private bool _hasUserSelectedDriverPackOption;
     private bool _isUpdatingFirmwareOptionSelection;
@@ -150,10 +107,6 @@ public partial class MainWindowViewModel : ObservableObject, IDisposable
     [ObservableProperty]
     [NotifyCanExecuteChangedFor(nameof(NextWizardStepCommand))]
     [NotifyCanExecuteChangedFor(nameof(StartDeploymentCommand))]
-    private OperatingSystemCatalogItem? selectedOperatingSystem;
-
-    [ObservableProperty]
-    [NotifyCanExecuteChangedFor(nameof(StartDeploymentCommand))]
     private TargetDiskInfo? selectedTargetDisk;
 
     [ObservableProperty]
@@ -197,36 +150,18 @@ public partial class MainWindowViewModel : ObservableObject, IDisposable
 
     [ObservableProperty]
     private string detectedHardwareSummary = "Detecting hardware...";
-
-    [ObservableProperty]
-    private string effectiveOsArchitecture = NormalizeArchitecture(Environment.GetEnvironmentVariable("PROCESSOR_ARCHITECTURE") ?? string.Empty);
-
-    [ObservableProperty]
-    private string selectedReleaseId = DefaultReleaseId;
-
-    [ObservableProperty]
-    private string selectedLanguageCode = DefaultLanguageCode;
-
-    [ObservableProperty]
-    private string selectedLicenseChannel = DefaultLicenseChannel;
-
-    [ObservableProperty]
-    private string selectedEdition = DefaultEdition;
-
-    public ObservableCollection<OperatingSystemCatalogItem> OperatingSystems { get; } = [];
-    public ObservableCollection<string> ReleaseIdFilters { get; } = [];
-    public ObservableCollection<string> LanguageFilters { get; } = [];
-    public ObservableCollection<string> LicenseChannelFilters { get; } = [];
-    public ObservableCollection<string> EditionFilters { get; } = [];
     public ObservableCollection<DriverPackCatalogItem> DriverPacks { get; } = [];
     public ObservableCollection<DriverPackOptionItem> DriverPackOptions { get; } = [];
     public ObservableCollection<string> DriverPackModelOptions { get; } = [];
     public ObservableCollection<string> DriverPackVersionOptions { get; } = [];
     public ObservableCollection<TargetDiskInfo> TargetDisks { get; } = [];
     public DeploymentSessionViewModel Session { get; }
+    public OperatingSystemCatalogViewModel OperatingSystemCatalog { get; }
 
     public DeployThemeMode CurrentTheme => _themeService.CurrentTheme;
     public bool IsDebugSafeMode => DebugSafetyMode.IsEnabled;
+    public string EffectiveOsArchitecture => OperatingSystemCatalog.EffectiveOsArchitecture;
+    public OperatingSystemCatalogItem? SelectedOperatingSystem => OperatingSystemCatalog.SelectedOperatingSystem;
     public string DriverPackModeDisplay => SelectedDriverPackOption?.Kind switch
     {
         DriverPackSelectionKind.MicrosoftUpdateCatalog => "Microsoft Update Catalog",
@@ -237,7 +172,6 @@ public partial class MainWindowViewModel : ObservableObject, IDisposable
     public bool IsDriverPackModelSelectionEnabled => IsOemDriverSourceSelected && DriverPackModelOptions.Count > 0;
     public bool IsDriverPackVersionSelectionEnabled => IsDriverPackModelSelectionEnabled && DriverPackVersionOptions.Count > 0;
     public bool IsFirmwareUpdatesOptionEnabled => _detectedHardware?.IsVirtualMachine != true;
-    public bool IsLanguageSelectionEnabled => !(_forceSingleVisibleLanguageSelection && LanguageFilters.Count == 1);
     public string SelectedDriverPackSelectionDisplay => BuildSelectedDriverPackSelectionDisplay();
     public bool HasTargetComputerNameValidationError => !string.IsNullOrWhiteSpace(TargetComputerNameValidationMessage);
     public string VersionDisplay => $"Version: {AppVersion}";
@@ -287,6 +221,10 @@ public partial class MainWindowViewModel : ObservableObject, IDisposable
         (DeploymentMode resolvedMode, string? resolvedUsbCacheRuntimeRoot) = ResolveDeploymentRuntimeContext();
         _resolvedDeploymentMode = resolvedMode;
         _resolvedUsbCacheRuntimeRoot = resolvedUsbCacheRuntimeRoot;
+        OperatingSystemCatalog = new OperatingSystemCatalogViewModel(
+            _logger,
+            Environment.GetEnvironmentVariable("PROCESSOR_ARCHITECTURE") ?? string.Empty);
+        OperatingSystemCatalog.StateChanged += OnOperatingSystemCatalogStateChanged;
         Session = new DeploymentSessionViewModel(
             _dispatcher,
             _logger,
@@ -340,19 +278,10 @@ public partial class MainWindowViewModel : ObservableObject, IDisposable
     private void ApplyExpertDeployConfiguration(FoundryDeployConfigurationDocument document)
     {
         ArgumentNullException.ThrowIfNull(document);
-
-        _configuredVisibleLanguageCodes.Clear();
-        foreach (string languageCode in document.Localization.VisibleLanguageCodes)
-        {
-            string normalized = NormalizeLanguageCode(languageCode);
-            if (!string.IsNullOrWhiteSpace(normalized))
-            {
-                _configuredVisibleLanguageCodes.Add(normalized);
-            }
-        }
-
-        _configuredDefaultLanguageCodeOverride = NormalizeOptionalLanguageCode(document.Localization.DefaultLanguageCodeOverride);
-        _forceSingleVisibleLanguageSelection = document.Localization.ForceSingleVisibleLanguage;
+        OperatingSystemCatalog.ApplyExpertLocalization(
+            document.Localization.VisibleLanguageCodes,
+            document.Localization.DefaultLanguageCodeOverride,
+            document.Localization.ForceSingleVisibleLanguage);
         _machineNamingConfiguration = document.Customization.MachineNaming ?? new DeployMachineNamingSettings();
         _lockedComputerNamePrefix = ComputerNameRules.Normalize(_machineNamingConfiguration.Prefix);
         IsTargetComputerNameReadOnly = _machineNamingConfiguration.IsEnabled && !_machineNamingConfiguration.AllowManualSuffixEdit;
@@ -364,8 +293,6 @@ public partial class MainWindowViewModel : ObservableObject, IDisposable
                 : TargetComputerName;
             ApplyManagedComputerNameValue(BuildConfiguredComputerName(seed));
         }
-
-        OnPropertyChanged(nameof(IsLanguageSelectionEnabled));
     }
 
     [RelayCommand]
@@ -410,23 +337,16 @@ public partial class MainWindowViewModel : ObservableObject, IDisposable
 
             RunOnUi(() =>
             {
-                OperatingSystems.Clear();
-                foreach (OperatingSystemCatalogItem item in operatingSystems)
-                {
-                    OperatingSystems.Add(item);
-                }
-
+                OperatingSystemCatalog.ApplyCatalog(operatingSystems);
                 DriverPacks.Clear();
                 foreach (DriverPackCatalogItem item in driverPacks)
                 {
                     DriverPacks.Add(item);
                 }
 
-                RefreshOsFilterOptions();
-                ApplyOsFilter();
                 RefreshDriverPackOptions();
 
-                Session.SetStatus($"Catalogs loaded: {OperatingSystems.Count} OS entries, {DriverPacks.Count} driver packs.");
+                Session.SetStatus($"Catalogs loaded: {OperatingSystemCatalog.OperatingSystems.Count} OS entries, {DriverPacks.Count} driver packs.");
             });
         }
         catch (Exception ex)
@@ -523,7 +443,7 @@ public partial class MainWindowViewModel : ObservableObject, IDisposable
     private async Task StartDeploymentAsync()
     {
         _logger.LogInformation("Start deployment requested.");
-        if (SelectedOperatingSystem is null)
+        if (OperatingSystemCatalog.SelectedOperatingSystem is null)
         {
             return;
         }
@@ -558,7 +478,7 @@ public partial class MainWindowViewModel : ObservableObject, IDisposable
             return;
         }
 
-        if (!IsDebugSafeMode && !ConfirmDestructiveDeployment(effectiveTargetDisk, SelectedOperatingSystem))
+        if (!IsDebugSafeMode && !ConfirmDestructiveDeployment(effectiveTargetDisk, OperatingSystemCatalog.SelectedOperatingSystem))
         {
             Session.SetStatus("Deployment cancelled by user.");
             return;
@@ -587,7 +507,7 @@ public partial class MainWindowViewModel : ObservableObject, IDisposable
             CacheRootPath = CacheRootPath,
             TargetDiskNumber = effectiveTargetDisk.DiskNumber,
             TargetComputerName = normalizedComputerName,
-            OperatingSystem = SelectedOperatingSystem,
+            OperatingSystem = OperatingSystemCatalog.SelectedOperatingSystem,
             DriverPackSelectionKind = effectiveDriverPackKind,
             DriverPack = effectiveDriverPack,
             ApplyFirmwareUpdates = ApplyFirmwareUpdates,
@@ -673,11 +593,6 @@ public partial class MainWindowViewModel : ObservableObject, IDisposable
             "Action: Verify disk attributes and retry deployment.");
     }
 
-    partial void OnSelectedOperatingSystemChanged(OperatingSystemCatalogItem? value)
-    {
-        RefreshDriverPackOptions();
-    }
-
     partial void OnTargetComputerNameChanged(string value)
     {
         if (_isApplyingManagedComputerName)
@@ -712,49 +627,6 @@ public partial class MainWindowViewModel : ObservableObject, IDisposable
         RefreshDriverPackVersionOptions();
     }
 
-    partial void OnEffectiveOsArchitectureChanged(string value)
-    {
-        if (_isUpdatingOsFilters)
-        {
-            return;
-        }
-
-        RefreshOsFilterOptions();
-        ApplyOsFilter();
-        RefreshDriverPackOptions();
-    }
-
-    partial void OnSelectedReleaseIdChanged(string value)
-    {
-        HandleOsFilterSelectionChanged();
-    }
-
-    partial void OnSelectedLanguageCodeChanged(string value)
-    {
-        HandleOsFilterSelectionChanged();
-    }
-
-    partial void OnSelectedLicenseChannelChanged(string value)
-    {
-        HandleOsFilterSelectionChanged();
-    }
-
-    partial void OnSelectedEditionChanged(string value)
-    {
-        HandleOsFilterSelectionChanged();
-    }
-
-    private void HandleOsFilterSelectionChanged()
-    {
-        if (_isUpdatingOsFilters)
-        {
-            return;
-        }
-
-        RefreshOsFilterOptions();
-        ApplyOsFilter();
-    }
-
     partial void OnAutoSelectDriverPackWhenEmptyChanged(bool value)
     {
         RefreshDriverPackOptions();
@@ -782,6 +654,15 @@ public partial class MainWindowViewModel : ObservableObject, IDisposable
         {
             Session.SetStatus($"Selected disk blocked: {value.SelectionWarning}");
         }
+    }
+
+    private void OnOperatingSystemCatalogStateChanged(object? sender, EventArgs e)
+    {
+        OnPropertyChanged(nameof(EffectiveOsArchitecture));
+        OnPropertyChanged(nameof(SelectedOperatingSystem));
+        RefreshDriverPackOptions();
+        NextWizardStepCommand.NotifyCanExecuteChanged();
+        StartDeploymentCommand.NotifyCanExecuteChanged();
     }
 
     private bool CanShowDebugPages()
@@ -839,12 +720,7 @@ public partial class MainWindowViewModel : ObservableObject, IDisposable
     private bool IsOsCatalogReadyForNavigation()
     {
         return !IsCatalogLoading &&
-               OperatingSystems.Count > 0 &&
-               ReleaseIdFilters.Count > 0 &&
-               LanguageFilters.Count > 0 &&
-               LicenseChannelFilters.Count > 0 &&
-               EditionFilters.Count > 0 &&
-               SelectedOperatingSystem is not null;
+               OperatingSystemCatalog.IsReadyForNavigation();
     }
 
     private bool CanStartDeployment()
@@ -861,7 +737,7 @@ public partial class MainWindowViewModel : ObservableObject, IDisposable
                !IsTargetDiskLoading &&
                WizardStepIndex == 3 &&
                ComputerNameRules.IsValid(TargetComputerName) &&
-               SelectedOperatingSystem is not null &&
+               OperatingSystemCatalog.SelectedOperatingSystem is not null &&
                hasTargetDisk &&
                HasValidDriverPackSelection();
     }
@@ -874,390 +750,6 @@ public partial class MainWindowViewModel : ObservableObject, IDisposable
         }
 
         return ResolveEffectiveDriverPackSelection() is not null;
-    }
-
-    private void ApplyOsFilter()
-    {
-        OperatingSystemCatalogItem[] filtered = BuildFilteredOperatingSystems();
-
-        OperatingSystemCatalogItem? matchingCurrent = SelectedOperatingSystem is null
-            ? null
-            : filtered.FirstOrDefault(item => IsSameOperatingSystemMedia(item, SelectedOperatingSystem));
-
-        OperatingSystemCatalogItem? selected = matchingCurrent ?? filtered.FirstOrDefault();
-        if (selected is null)
-        {
-            SelectedOperatingSystem = null;
-            return;
-        }
-
-        SelectedOperatingSystem = ApplyEditionSelection(selected);
-    }
-
-    private void RefreshOsFilterOptions()
-    {
-        _isUpdatingOsFilters = true;
-
-        try
-        {
-            string previousReleaseId = SelectedReleaseId;
-            string previousLanguageCode = SelectedLanguageCode;
-            string previousLicenseChannel = SelectedLicenseChannel;
-            string previousEdition = SelectedEdition;
-
-            IEnumerable<OperatingSystemCatalogItem> baseQuery = BuildOsQueryWithArchitecture(OperatingSystems);
-
-            IEnumerable<OperatingSystemCatalogItem> releaseScope = baseQuery;
-            SelectedReleaseId = UpdateFilterSelection(
-                ReleaseIdFilters,
-                releaseScope.Select(item => item.ReleaseId),
-                previousReleaseId,
-                DefaultReleaseId,
-                selectFirstWhenNoMatch: true);
-
-            IEnumerable<OperatingSystemCatalogItem> languageScope = ApplyReleaseIdFilter(releaseScope);
-            IEnumerable<string> effectiveLanguageValues = BuildEffectiveLanguageFilterValues(languageScope);
-            SelectedLanguageCode = UpdateLanguageFilterSelection(
-                LanguageFilters,
-                effectiveLanguageValues,
-                previousLanguageCode);
-
-            string configuredDefaultLanguageCode = EnsureLanguageSelection(
-                _configuredDefaultLanguageCodeOverride ?? string.Empty,
-                LanguageFilters);
-            if (!string.IsNullOrWhiteSpace(configuredDefaultLanguageCode))
-            {
-                SelectedLanguageCode = configuredDefaultLanguageCode;
-            }
-            else if (!_hasLoggedUnavailableDefaultLanguageOverride &&
-                     !string.IsNullOrWhiteSpace(_configuredDefaultLanguageCodeOverride))
-            {
-                _logger.LogWarning(
-                    "Configured default language override '{LanguageCode}' is not available in the current catalog scope and was ignored.",
-                    _configuredDefaultLanguageCodeOverride);
-                _hasLoggedUnavailableDefaultLanguageOverride = true;
-            }
-
-            IEnumerable<OperatingSystemCatalogItem> licenseScope = ApplyLanguageFilter(languageScope);
-            SelectedLicenseChannel = UpdateFilterSelection(
-                LicenseChannelFilters,
-                licenseScope.Select(item => item.LicenseChannel),
-                previousLicenseChannel,
-                DefaultLicenseChannel,
-                selectFirstWhenNoMatch: true);
-
-            IEnumerable<OperatingSystemCatalogItem> editionScope = ApplyLicenseChannelFilter(licenseScope);
-            IEnumerable<string> recommendedEditions = BuildRecommendedEditionOptions(editionScope);
-            SelectedEdition = UpdateFilterSelection(
-                EditionFilters,
-                recommendedEditions,
-                previousEdition,
-                DefaultEdition,
-                selectFirstWhenNoMatch: true);
-        }
-        finally
-        {
-            _isUpdatingOsFilters = false;
-            OnPropertyChanged(nameof(IsLanguageSelectionEnabled));
-        }
-    }
-
-    private IEnumerable<OperatingSystemCatalogItem> BuildOsQueryWithArchitecture(IEnumerable<OperatingSystemCatalogItem> source)
-    {
-        IEnumerable<OperatingSystemCatalogItem> supportedSource = source.Where(OperatingSystemSupportMatrix.IsSupported);
-        string architecture = NormalizeArchitecture(EffectiveOsArchitecture);
-        if (string.IsNullOrWhiteSpace(architecture))
-        {
-            return supportedSource;
-        }
-
-        return supportedSource.Where(item => IsArchitectureMatch(item.Architecture, architecture));
-    }
-
-    private IEnumerable<OperatingSystemCatalogItem> ApplyReleaseIdFilter(IEnumerable<OperatingSystemCatalogItem> source)
-    {
-        return IsFilterUnset(SelectedReleaseId)
-            ? source
-            : source.Where(item => item.ReleaseId.Equals(SelectedReleaseId, StringComparison.OrdinalIgnoreCase));
-    }
-
-    private IEnumerable<OperatingSystemCatalogItem> ApplyLanguageFilter(IEnumerable<OperatingSystemCatalogItem> source)
-    {
-        return IsFilterUnset(SelectedLanguageCode)
-            ? source
-            : source.Where(item => GetLanguageFilterValue(item).Equals(SelectedLanguageCode, StringComparison.OrdinalIgnoreCase));
-    }
-
-    private IEnumerable<string> BuildEffectiveLanguageFilterValues(IEnumerable<OperatingSystemCatalogItem> source)
-    {
-        string[] availableLanguageValues = source
-            .Select(GetLanguageFilterValue)
-            .Where(value => !string.IsNullOrWhiteSpace(value))
-            .Distinct(StringComparer.OrdinalIgnoreCase)
-            .ToArray();
-
-        if (_configuredVisibleLanguageCodes.Count == 0)
-        {
-            return availableLanguageValues;
-        }
-
-        string[] filteredLanguageValues = availableLanguageValues
-            .Where(value => _configuredVisibleLanguageCodes.Contains(NormalizeLanguageCode(value)))
-            .ToArray();
-
-        if (filteredLanguageValues.Length > 0)
-        {
-            return filteredLanguageValues;
-        }
-
-        if (!_hasLoggedUnavailableConfiguredLanguages)
-        {
-            string configuredLanguages = string.Join(", ", _configuredVisibleLanguageCodes.OrderBy(value => value, StringComparer.OrdinalIgnoreCase));
-            _logger.LogWarning(
-                "Configured visible languages [{ConfiguredLanguages}] do not match the current catalog scope. Falling back to the catalog languages.",
-                configuredLanguages);
-            _hasLoggedUnavailableConfiguredLanguages = true;
-        }
-
-        return availableLanguageValues;
-    }
-
-    private IEnumerable<OperatingSystemCatalogItem> ApplyLicenseChannelFilter(IEnumerable<OperatingSystemCatalogItem> source)
-    {
-        return IsFilterUnset(SelectedLicenseChannel)
-            ? source
-            : source.Where(item => item.LicenseChannel.Equals(SelectedLicenseChannel, StringComparison.OrdinalIgnoreCase));
-    }
-
-    private static bool IsFilterUnset(string value)
-    {
-        return string.IsNullOrWhiteSpace(value);
-    }
-
-    private static string GetLanguageFilterValue(OperatingSystemCatalogItem item)
-    {
-        return !string.IsNullOrWhiteSpace(item.LanguageCode)
-            ? item.LanguageCode
-            : item.Language;
-    }
-
-    private IEnumerable<string> BuildRecommendedEditionOptions(IEnumerable<OperatingSystemCatalogItem> scope)
-    {
-        string architecture = NormalizeArchitecture(EffectiveOsArchitecture);
-        bool hasRetail = scope.Any(item => item.LicenseChannel.Equals("RET", StringComparison.OrdinalIgnoreCase));
-        bool hasVolume = scope.Any(item => item.LicenseChannel.Equals("VOL", StringComparison.OrdinalIgnoreCase));
-
-        List<string> recommended = [];
-
-        if (IsFilterUnset(SelectedLicenseChannel))
-        {
-            if (hasRetail)
-            {
-                recommended.AddRange(GetEditionOptionsForChannel("RET", architecture));
-            }
-
-            if (hasVolume)
-            {
-                recommended.AddRange(GetEditionOptionsForChannel("VOL", architecture));
-            }
-        }
-        else
-        {
-            recommended.AddRange(GetEditionOptionsForChannel(SelectedLicenseChannel, architecture));
-        }
-
-        // Fall back to raw catalog editions if no curated mapping applies.
-        if (recommended.Count == 0)
-        {
-            recommended.AddRange(scope.Select(item => item.Edition));
-        }
-
-        return recommended;
-    }
-
-    private static IEnumerable<string> GetEditionOptionsForChannel(string licenseChannel, string architecture)
-    {
-        bool isArm64 = architecture.Equals("arm64", StringComparison.OrdinalIgnoreCase);
-
-        if (licenseChannel.Equals("VOL", StringComparison.OrdinalIgnoreCase))
-        {
-            return isArm64 ? Arm64VolumeEditionOptions : VolumeEditionOptions;
-        }
-
-        if (licenseChannel.Equals("RET", StringComparison.OrdinalIgnoreCase))
-        {
-            return isArm64 ? Arm64RetailEditionOptions : RetailEditionOptions;
-        }
-
-        return [];
-    }
-
-    private static void UpdateFilterCollection(ObservableCollection<string> target, IEnumerable<string> values)
-    {
-        string[] normalizedValues = values
-            .Where(value => !string.IsNullOrWhiteSpace(value))
-            .Distinct(StringComparer.OrdinalIgnoreCase)
-            .OrderBy(value => value, StringComparer.OrdinalIgnoreCase)
-            .ToArray();
-
-        target.Clear();
-
-        foreach (string value in normalizedValues)
-        {
-            target.Add(value);
-        }
-    }
-
-    private static bool TryGetFilterSelection(string? selectedValue, ObservableCollection<string> options, out string matched)
-    {
-        matched = string.Empty;
-
-        if (options.Count == 0 || IsFilterUnset(selectedValue ?? string.Empty))
-        {
-            return false;
-        }
-
-        string? matchingOption = options.FirstOrDefault(option =>
-            option.Equals(selectedValue, StringComparison.OrdinalIgnoreCase));
-
-        if (string.IsNullOrWhiteSpace(matchingOption))
-        {
-            return false;
-        }
-
-        matched = matchingOption;
-        return true;
-    }
-
-    private static string UpdateFilterSelection(
-        ObservableCollection<string> target,
-        IEnumerable<string> values,
-        string previousSelection,
-        string? defaultSelection = null,
-        bool selectFirstWhenNoMatch = false)
-    {
-        UpdateFilterCollection(target, values);
-
-        if (target.Count == 0)
-        {
-            return string.Empty;
-        }
-
-        if (TryGetFilterSelection(previousSelection, target, out string selected))
-        {
-            return selected;
-        }
-
-        if (!string.IsNullOrWhiteSpace(defaultSelection) &&
-            TryGetFilterSelection(defaultSelection, target, out selected))
-        {
-            return selected;
-        }
-
-        return selectFirstWhenNoMatch ? target[0] : string.Empty;
-    }
-
-    private static string UpdateLanguageFilterSelection(
-        ObservableCollection<string> target,
-        IEnumerable<string> values,
-        string previousSelection)
-    {
-        UpdateFilterCollection(target, values);
-        if (target.Count == 0)
-        {
-            return string.Empty;
-        }
-
-        foreach (string candidate in new[] { previousSelection, DefaultLanguageCode, FallbackLanguageCode })
-        {
-            string selected = EnsureLanguageSelection(candidate, target);
-            if (!string.IsNullOrWhiteSpace(selected))
-            {
-                return selected;
-            }
-        }
-
-        return target[0];
-    }
-
-    private static string EnsureLanguageSelection(string languageCode, ObservableCollection<string> options)
-    {
-        if (options.Count == 0)
-        {
-            return string.Empty;
-        }
-
-        if (IsFilterUnset(languageCode))
-        {
-            return string.Empty;
-        }
-
-        string normalized = NormalizeLanguageCode(languageCode);
-
-        string? exact = options.FirstOrDefault(option =>
-            NormalizeLanguageCode(option).Equals(normalized, StringComparison.OrdinalIgnoreCase));
-
-        if (!string.IsNullOrWhiteSpace(exact))
-        {
-            return exact;
-        }
-
-        string neutral = normalized.Split('-', 2, StringSplitOptions.RemoveEmptyEntries)[0];
-        if (!string.IsNullOrWhiteSpace(neutral))
-        {
-            string? sameLanguage = options.FirstOrDefault(option =>
-            {
-                string candidate = NormalizeLanguageCode(option);
-                return candidate.Equals(neutral, StringComparison.OrdinalIgnoreCase) ||
-                       candidate.StartsWith($"{neutral}-", StringComparison.OrdinalIgnoreCase);
-            });
-
-            if (!string.IsNullOrWhiteSpace(sameLanguage))
-            {
-                return sameLanguage;
-            }
-        }
-
-        return string.Empty;
-    }
-
-    private static string ResolveDefaultLanguageCode()
-    {
-        string[] candidates =
-        [
-            CultureInfo.CurrentUICulture.Name,
-            CultureInfo.CurrentCulture.Name,
-            CultureInfo.InstalledUICulture.Name
-        ];
-
-        foreach (string candidate in candidates)
-        {
-            string normalized = NormalizeLanguageCode(candidate);
-            if (!string.IsNullOrWhiteSpace(normalized))
-            {
-                return normalized;
-            }
-        }
-
-        return FallbackLanguageCode;
-    }
-
-    private static string? NormalizeOptionalLanguageCode(string? languageCode)
-    {
-        if (string.IsNullOrWhiteSpace(languageCode))
-        {
-            return null;
-        }
-
-        string normalized = NormalizeLanguageCode(languageCode);
-        return string.IsNullOrWhiteSpace(normalized)
-            ? null
-            : normalized;
-    }
-
-    private static string NormalizeLanguageCode(string languageCode)
-    {
-        return languageCode.Trim().Replace('_', '-').ToLowerInvariant();
     }
 
     private string NormalizeManagedComputerNameValue(string? value)
@@ -1910,21 +1402,6 @@ public partial class MainWindowViewModel : ObservableObject, IDisposable
             .ToArray();
     }
 
-    private OperatingSystemCatalogItem[] BuildFilteredOperatingSystems()
-    {
-        IEnumerable<OperatingSystemCatalogItem> query = BuildOsQueryWithArchitecture(OperatingSystems);
-        query = ApplyReleaseIdFilter(query);
-        query = ApplyLanguageFilter(query);
-        query = ApplyLicenseChannelFilter(query);
-
-        return query
-            .GroupBy(
-                item => string.IsNullOrWhiteSpace(item.Url) ? item.FileName : item.Url,
-                StringComparer.OrdinalIgnoreCase)
-            .Select(group => group.First())
-            .ToArray();
-    }
-
     private async Task LoadHardwareProfileAsync()
     {
         try
@@ -1933,12 +1410,10 @@ public partial class MainWindowViewModel : ObservableObject, IDisposable
             RunOnUi(() =>
             {
                 _detectedHardware = profile;
-                EffectiveOsArchitecture = NormalizeArchitecture(profile.Architecture);
+                OperatingSystemCatalog.SetEffectiveArchitecture(profile.Architecture);
                 SyncFirmwareOptionFromHardware(profile);
                 DetectedHardwareSummary =
                     $"{profile.DisplayLabel} | TPM: {(profile.IsTpmPresent ? "Yes" : "No")} | Autopilot: {(profile.IsAutopilotCapable ? "Capable" : "Needs checks")} | Power: {(profile.IsOnBattery ? "Battery" : "AC")} | Firmware: {(profile.SystemFirmwareHardwareId.Length > 0 ? "Detected" : "Unavailable")}";
-                RefreshOsFilterOptions();
-                ApplyOsFilter();
                 RefreshDriverPackOptions();
             });
             _logger.LogInformation("Hardware profile loaded in view model. DisplayLabel={DisplayLabel}", profile.DisplayLabel);
@@ -1976,28 +1451,6 @@ public partial class MainWindowViewModel : ObservableObject, IDisposable
         }
 
         OnPropertyChanged(nameof(IsFirmwareUpdatesOptionEnabled));
-    }
-
-    private OperatingSystemCatalogItem ApplyEditionSelection(OperatingSystemCatalogItem item)
-    {
-        if (IsFilterUnset(SelectedEdition))
-        {
-            return item;
-        }
-
-        if (item.Edition.Equals(SelectedEdition, StringComparison.OrdinalIgnoreCase))
-        {
-            return item;
-        }
-
-        return item with { Edition = SelectedEdition };
-    }
-
-    private static bool IsSameOperatingSystemMedia(OperatingSystemCatalogItem left, OperatingSystemCatalogItem right)
-    {
-        string leftKey = string.IsNullOrWhiteSpace(left.Url) ? left.FileName : left.Url;
-        string rightKey = string.IsNullOrWhiteSpace(right.Url) ? right.FileName : right.Url;
-        return leftKey.Equals(rightKey, StringComparison.OrdinalIgnoreCase);
     }
 
     private async Task LoadOfflineComputerNameAsync()
@@ -2204,6 +1657,7 @@ public partial class MainWindowViewModel : ObservableObject, IDisposable
             return;
         }
 
+        OperatingSystemCatalog.StateChanged -= OnOperatingSystemCatalogStateChanged;
         Session.Dispose();
         _isDisposed = true;
     }
